requirements.txt
psycopg2-binary==2.9.9
pyyaml==6.0.1
pytz==2024.1
tqdm==4.66.4
sqlparse==0.5.0
#pysqlite3==0.5.2
mysql-connector-python==8.4.0
boto3==1.34.112